requirements.txt
Flask
python-dotenv
requests
feedparser
bunnet
pymongo
